API_serving/requirements.txt
ultralytics
opencv-python-headless
sahi

onnx==1.12.0
onnxruntime-gpu==1.12.1

Flask==2.2.3
apiflask
flask-cors

gunicorn==20.1.0
json2html

# Laser
albumentations==1.3.1
anomalib==0.7.0
openvino==2023.0.2
